+++ pisek/__main__.py
@@ -21,10 +21,10 @@ import sys
 import signal
 
 from pisek.jobs.task_pipeline import TaskPipeline
-from pisek.pipeline_tools import run_pipeline, load_env, PATH, locked_folder
+from pisek.utils.pipeline_tools import run_pipeline, PATH, locked_folder
 
-from pisek.util import clean_task_dir
-from pisek.terminal import eprint
+from pisek.utils.util import clean_task_dir
+from pisek.utils.terminal import eprint
 from pisek.license import license, license_gnu
 import pisek.cms as cms
 from pisek.update_config import update
