+++ pisek/cms/submit.py
@@ -2,7 +2,7 @@ import subprocess
 import glob
 import os
 import tempfile
-import pisek.util as util
+import pisek.utils.util as util
 from pisek.config.task_config import TaskConfig
 from . import check
