+++ pisek/jobs/parts/checker.py
@@ -20,7 +20,7 @@ from pisek.jobs.jobs import State, Job, PipelineItemFailure
 from pisek.config.env import Env
 from pisek.paths import TaskPath
 from pisek.config.task_config import ProgramType
-from pisek.terminal import colored
+from pisek.utils.terminal import colored
 from pisek.jobs.parts.task_job import TaskJobManager
 from pisek.jobs.parts.program import RunResult, RunResultKind, ProgramsJob
 from pisek.jobs.parts.compile import Compile
