+++ pisek/jobs/parts/program.py
@@ -28,10 +28,9 @@ from pisek.config.task_config import ProgramType
 from pisek.config.env import Env
 from pisek.paths import TaskPath
 from pisek.jobs.jobs import PipelineItemFailure
-from pisek.terminal import tab, colored
+from pisek.utils.text import tab
+from pisek.utils.terminal import colored
 from pisek.jobs.parts.task_job import TaskHelper, TaskJob
-
-import pisek.util as util
 
 
 class RunResultKind(Enum):
