+++ pisek/jobs/parts/solution.py
@@ -19,12 +19,12 @@ import tempfile
 import time
 from typing import Any, Optional, Callable, Iterable
 
-import pisek.util as util
 from pisek.jobs.jobs import State, Job, PipelineItemFailure
 from pisek.config.env import Env
 from pisek.paths import TaskPath
 from pisek.config.task_config import ProgramType
-from pisek.terminal import pad, pad_left, tab, MSG_LEN
+from pisek.utils.text import pad, pad_left, tab
+from pisek.utils.terminal import MSG_LEN
 from pisek.jobs.status import MAX_BAR_LEN
 from pisek.jobs.parts.task_job import TaskJobManager
 from pisek.jobs.parts.program import RunResult, ProgramsJob
