+++ pisek/visualize.py
@@ -24,7 +24,6 @@ import sys
 from colorama import Fore
 from typing import Optional, Union, Iterable, Callable
 
-from pisek import util
 from .config.task_config import load_config, TaskConfig, SubtaskConfig
 from pisek.jobs.parts.task_job import TaskHelper
